Moved lightmapper from /scene to /framework
- majority of Lightmapper has move to framework, as it uses entities / components and so that’s where it belongs
- LightmapCache stayed behind on the scene level, as mesh instances need to communicate with it
