+++ src/framework/app-options.js
@@ -6,7 +6,7 @@
 /** @typedef {import('../platform/input/mouse.js').Mouse} Mouse */
 /** @typedef {import('../platform/input/touch-device.js').TouchDevice} TouchDevice */
 /** @typedef {import('../platform/sound/manager.js').SoundManager} SoundManager */
-/** @typedef {import('../scene/lightmapper/lightmapper.js').Lightmapper} Lightmapper */
+/** @typedef {import('./lightmapper/lightmapper.js').Lightmapper} Lightmapper */
 /** @typedef {import('../scene/batching/batch-manager.js').BatchManager} BatchManager */
 /** @typedef {import('./components/system.js').ComponentSystem} ComponentSystem */
 /** @typedef {import('./xr/xr-manager.js').XrManager} XrManager */
