+++ src/framework/lightmapper/bake-light.js
@@ -1,6 +1,6 @@
 import { BoundingBox } from '../../core/shape/bounding-box.js';
 import { BoundingSphere } from '../../core/shape/bounding-sphere.js';
-import { LIGHTTYPE_DIRECTIONAL } from '../constants.js';
+import { LIGHTTYPE_DIRECTIONAL } from '../../scene/constants.js';
 
 const tempSphere = new BoundingSphere();
