+++ src/framework/lightmapper/bake-mesh-node.js
