+++ src/scene/graphics/lightmap-cache.js
